OWRS-299 #time 1h testing optical flow

+++ rover/src/owr_positioning/include/optical_localization.h
@@ -18,7 +18,7 @@ class optical_localization {
 		
 		ros::Time prev_time;
 		
-		const unsigned int pixels_per_metre_traversed;
+		const double pixels_per_metre_traversed;
 		const double linear_velocity_scale;
 		const double rotation_velocity_scale;
